requirements.txt
torch
torchvision
pandas
numpy
pydicom
Pillow
scikit-learn
tqdm
